feat: 1.优化按照分类查询 2.解决商品明细 规格信息异常

+++ spring-cloud-product/src/main/java/com/zhangz/demo/spring/cloud/product/dao/GoodPropertyMapper.java
@@ -11,6 +11,6 @@ import java.util.Set;
 @Mapper
 public interface GoodPropertyMapper extends BaseMapper<GoodProperty> {
 
-    Set<GoodProperty> queryByPropertyId(@Param("propertyId") int propertyId);
+    Set<GoodProperty> queryByPropertyId(@Param("propertyId") long propertyId);
     
 }
